Add files via upload
In this project, we will develop and evaluate the performance and the predictive power of a model trained and tested on data collected from houses in Boston’s suburbs. Once we get a good fit, we will use this model to predict the monetary value of a house located at the Boston’s area. 

Our objective is to create a Machine Learning Model to Learn and predict the price of a house using Boston Housing Dataset.